One file declaration per mutator
